Updates to include fannie non-standard (exclusion) loans

+++ mortgage_imports/fannie.py
@@ -2,7 +2,7 @@ import pkg_resources
 import mortgage_imports.clickhouse_utilities as cu
 import os
 
-def load_fannie(data_loc):
+def load_fannie(data_loc: str, src: str, first: bool):
     client = cu.make_connection()
     sql_loc = pkg_resources.resource_filename('mortgage_imports', 'sql/fannie') + '/'
     
@@ -12,16 +12,36 @@ def load_fannie(data_loc):
     # create DB if not there
     cu.run_query("CREATE DATABASE IF NOT EXISTS fannie", client)
 
-    # load harp_map table which maps non-harp loans to their harp refis
-    cu.run_query("DROP TABLE IF EXISTS fannie.harp_map", client)
-    cu.run_query(sql_loc + "harp_map_ct.sql", client, True)
-    cu.import_flat_file("fannie.harp_map", data_loc + "Loan_Mapping.txt", delim=",")
+    if src == 'standard':
+        # load harp_map table which maps non-harp loans to their harp refis
+        cu.run_query("DROP TABLE IF EXISTS fannie.harp_map", client)
+        cu.run_query(sql_loc + "harp_map_ct.sql", client, True)
+        cu.import_flat_file("fannie.harp_map", data_loc + "Loan_Mapping.txt", delim=",")
+        extra_fields=''
+        extra_field_ins = """
+            'N' AS ln_nonstd_doc_flg,
+            'N' AS ln_nonstd_uw_flg,
+            'N' AS ln_govt_guar_flg,
+            'N' AS ln_negam_flg
+        """
+    else:
+        extra_fields = """
+           ,ln_nonstd_doc_flg                 LowCardinality(FixedString(1)),
+            ln_nonstd_uw_flg                  LowCardinality(FixedString(1)),
+            ln_govt_guar_flg                  LowCardinality(FixedString(1)),
+            ln_negam_flg                      LowCardinality(FixedString(1))
+        """
+        extra_field_ins = """
+            ln_nonstd_doc_flg,
+            ln_nonstd_uw_flg,
+            ln_govt_guar_flg,
+            ln_negam_flg
+        """
 
     # files in the data directory
     flist = os.listdir(data_loc)
 
     # need to know if this is the first time through the loop
-    first = True
     # file count
     fn = 0
 
@@ -31,12 +51,16 @@ def load_fannie(data_loc):
             print("working on {0}".format(filename))
             src_file = filename[0:filename.find(".")]
             cu.run_query("DROP TABLE IF EXISTS fannie.raw", client)
-            cu.run_query(sql_loc + "raw_ct.sql", client, True)
+            cu.run_query(sql_loc + "raw_ct.sql", client, True,
+                         replace_source=['<extra_fields>'],
+                         replace_dest=[extra_fields])
             cu.import_flat_file("fannie.raw", data_loc + filename)
         
             cu.run_query("DROP TABLE IF EXISTS fannie.trans", client)
             cu.run_query(sql_loc + "transform_ct.sql", client, True)
-            cu.run_query(sql_loc + "transform_ins.sql", client, True)
+            cu.run_query(sql_loc + "transform_ins.sql", client, True,
+                         replace_source=['<src>', '<extra_field_ins>'],
+                         replace_dest=[src, extra_field_ins])
             cu.run_query("DROP TABLE IF EXISTS fannie.raw", client)
         
             cu.run_query("DROP TABLE IF EXISTS fannie.with_hpi", client)
@@ -46,7 +70,7 @@ def load_fannie(data_loc):
         
             cu.run_query("DROP TABLE IF EXISTS fannie.n3sted", client)
             cu.run_query(sql_loc + "nested_ct.sql", client, True)
-            cu.run_query(sql_loc + "nested_ins.sql", client, True, "XXXXXX", src_file)
+            cu.run_query(sql_loc + "nested_ins.sql", client, True, '<src_file>', src_file)
             cu.run_query("DROP TABLE IF EXISTS fannie.with_hpi", client)
             if first:
                 first = False
